+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `SliderValue::single` and `SliderValue::planar` helpers.
 - Rename `AnimatableAttribute` to `AnimatedAttribute`.
+- Add derives for setting up animated components and newtype components: `StaticComponent`, `ResponsiveComponent`, `AnimatedComponent`, `StaticReactComponent`, `ResponsiveReactComponent`, `AnimatedReactComponent`, `StaticNewtype`, `ResponsiveNewtype`, `AnimatedNewtype`, `StaticReactNewtype`, `ResponsiveReactNewtype`, `AnimatedReactNewtype`.
 
 ## 0.5.0